refactor(routes): generic types & file structure
